simplify projectors and measurements
